# Library-Manager/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Library Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /movies:
    get:
      operationId: listMovies
      tags: [movies]
      summary: List all movies
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: genre
          in: query
          schema:
            type: string
        - name: available
          in: query
          schema:
            type: boolean
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of movies
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MovieListResponse"

    post:
      operationId: createMovie
      tags: [movies]
      summary: Create a new movie
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMovieBody"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"
  /movies/{id}:
    get:
      operationId: getMovie
      tags: [movies]
      summary: Get a movie
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Movie details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"

    put:
      operationId: updateMovie
      tags: [movies]
      summary: Update a movie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMovieBody"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Movie"

    delete:
      operationId: deleteMovie
      tags: [movies]
      summary: Delete a movie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /transactions:
    get:
      operationId: listTransactions
      tags: [transactions]
      summary: List all transactions
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, issued, returned, overdue]
        - name: userId
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"

    post:
      operationId: createTransaction
      tags: [transactions]
      summary: Issue a book or movie
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionBody"
      responses:
        "201":
          description: Transaction created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}:
    get:
      operationId: getTransaction
      tags: [transactions]
      summary: Get a transaction
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Transaction details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/return:
    post:
      operationId: returnItem
      tags: [transactions]
      summary: Return a book or movie
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Returned
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/pay-fine:
    post:
      operationId: payFine
      tags: [transactions]
      summary: Pay the fine for an overdue item
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Fine paid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /transactions/{id}/approve:
    post:
      operationId: approveTransaction
      tags: [transactions]
      summary: Approve a pending issue request
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Movie:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        director:
          type: string
        genre:
          type: string
        description:
          type: string
        duration:
          type: integer
        releaseYear:
          type: integer
        totalCopies:
          type: integer
        availableCopies:
          type: integer
        status:
          type: string
          enum: [available, issued, unavailable]
        posterUrl:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - title
        - director
        - genre
        - totalCopies
        - availableCopies
        - status
    CreateMovieBody:
      type: object
      properties:
        title:
          type: string
        director:
          type: string
        genre:
          type: string
        description:
          type: string
        duration:
          type: integer
        releaseYear:
          type: integer
        totalCopies:
          type: integer
        posterUrl:
          type: string
      required:
        - title
        - director
        - genre
        - totalCopies
    MovieListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Movie"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...
    Transaction:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        itemType:
          type: string
          enum: [book, movie]
        itemId:
          type: integer
        itemTitle:
          type: string
        itemAuthor:
          type: string
        status:
          type: string
          enum: [pending, issued, returned, overdue]
        issueDate:
          type: string
          format: date-time
        dueDate:
          type: string
          format: date-time
        returnDate:
          type: string
          format: date-time
        fineAmount:
          type: number
        finePaid:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - userId
        - itemType
        - itemId
        - itemTitle
        - status
        - dueDate
    CreateTransactionBody:
      type: object
      properties:
        itemType:
          type: string
          enum: [book, movie]
        itemId:
          type: integer
        userId:
          type: integer
        daysToReturn:
          type: integer
          default: 14
      required:
        - itemType
        - itemId
    TransactionListResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit
...