# Library-Manager/lib/api-zod/src/generated/types/createMovieBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface CreateMovieBody {
  title: string;
  director: string;
  genre: string;
  description?: string;
  duration?: number;
  releaseYear?: number;
  totalCopies: number;
  posterUrl?: string;
}