# Library-Manager/lib/api-zod/src/generated/types/movieListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Library Management System API
 * OpenAPI spec version: 0.1.0
 */
import type { Movie } from "./movie";

export interface MovieListResponse {
  data: Movie[];
  total: number;
  page: number;
  limit: number;
}